# Optimize uuid.String