--- a/uuid/uuid.go
+++ b/uuid/uuid.go
@@ -4,7 +4,7 @@ package uuid
 
 import (
 	"crypto/rand"
-	"fmt"
+	"encoding/hex"
 )
 
 type UUID [16]byte
@@ -23,5 +23,17 @@ func NewV4() *UUID {
 }
 
 func (u *UUID) String() string {
-	return fmt.Sprintf("%x-%x-%x-%x-%x", u[:4], u[4:6], u[6:8], u[8:10], u[10:])
+	buf := make([]byte, 36)
+
+	hex.Encode(buf[0:8], u[0:4])
+	buf[8] = '-'
+	hex.Encode(buf[9:13], u[4:6])
+	buf[13] = '-'
+	hex.Encode(buf[14:18], u[6:8])
+	buf[18] = '-'
+	hex.Encode(buf[19:23], u[8:10])
+	buf[23] = '-'
+	hex.Encode(buf[24:], u[10:])
+
+	return string(buf)
 }