Write hex encoded data to a buffer instead of using fmt.Sprintf.

```
name     old time/op    new time/op    delta
String-2     412ns ± 5%      68ns ±14%  -83.41%  (p=0.000 n=9+9)

name     old alloc/op   new alloc/op   delta
String-2      208B ± 0%       48B ± 0%  -76.92%  (p=0.000 n=10+10)

name     old allocs/op  new allocs/op  delta
String-2      6.00 ± 0%      1.00 ± 0%  -83.33%  (p=0.000 n=10+10)
```